+++ clippinator/__main__.py
@@ -4,9 +4,9 @@ import rich
 import typer
 from dotenv import load_dotenv
 
-from clippy.minions.taskmaster import Taskmaster
-from clippy.project import Project
-from clippy.tools.utils import text_prompt
+from clippinator.minions.taskmaster import Taskmaster
+from clippinator.project import Project
+from clippinator.tools.utils import text_prompt
 
 load_dotenv()
 
@@ -16,20 +16,20 @@ app = typer.Typer(help="Clippy is an AI coding assistant.")
 @app.command()
 def taskmaster(project_path: str, objective: str = ""):
     """
-    Create a new project using clippy.
+    Create a new project using clippinator.
     """
     try:
         if not objective and not os.path.exists(
-                os.path.join(project_path, ".clippy.pkl")
+                os.path.join(project_path, ".clippinator.pkl")
         ):
             objective = text_prompt("What project do I need to create?\n")
-        if not objective and os.path.exists(os.path.join(project_path, ".clippy.pkl")):
-            print(os.path.join(project_path, ".clippy.pkl"))
-            tm = Taskmaster.load_from_file(os.path.join(project_path, ".clippy.pkl"))
+        if not objective and os.path.exists(os.path.join(project_path, ".clippinator.pkl")):
+            print(os.path.join(project_path, ".clippinator.pkl"))
+            tm = Taskmaster.load_from_file(os.path.join(project_path, ".clippinator.pkl"))
             tm.run(**tm.project.prompt_fields())
             return
-        elif os.path.exists(os.path.join(project_path, ".clippy.pkl")):
-            tm = Taskmaster.load_from_file(os.path.join(project_path, ".clippy.pkl"))
+        elif os.path.exists(os.path.join(project_path, ".clippinator.pkl")):
+            tm = Taskmaster.load_from_file(os.path.join(project_path, ".clippinator.pkl"))
             project = tm.project
             project.objective = objective
             tm = Taskmaster(project)
