+++ clippinator/minions/__init__.py
