+++ clippinator/minions/memory.py
