+++ clippinator/minions/prompts.py
