+++ clippinator/minions/specialized_minions.yaml
