+++ clippinator/project/project.py
@@ -4,7 +4,7 @@ import os
 import subprocess
 from dataclasses import dataclass, field
 
-from clippy.project.project_summary import get_file_summary
+from clippinator.project.project_summary import get_file_summary
 
 
 @dataclass
@@ -36,7 +36,7 @@ class Project:
             dir2:
                 file3.py
         """
-        from clippy.tools.utils import skip_file, skip_file_summary, trim_extra
+        from clippinator.tools.utils import skip_file, skip_file_summary, trim_extra
 
         res = ""
         if not os.path.isdir(path):
@@ -65,8 +65,8 @@ class Project:
         return res
 
     def lint(self, path: str = ''):
-        from clippy.tools.code_tools import lint_project
-        from clippy.tools.utils import trim_extra
+        from clippinator.tools.code_tools import lint_project
+        from clippinator.tools.utils import trim_extra
 
         path = os.path.join(self.path, path)
         path = path or self.path
@@ -81,8 +81,8 @@ class Project:
         return lint_project(path)
 
     def lint_file(self, path: str):
-        from clippy.tools.code_tools import lint_file
-        from clippy.tools.utils import trim_extra
+        from clippinator.tools.code_tools import lint_file
+        from clippinator.tools.utils import trim_extra
 
         path = os.path.join(self.path, path)
         if self.ci_commands.get('lintfile', '').strip():
@@ -101,7 +101,7 @@ class Project:
         return self.summary_cache
 
     def menu(self, prompt=None):
-        from clippy.tools.utils import select, get_input_from_editor
+        from clippinator.tools.utils import select, get_input_from_editor
         prompt_options = ["Edit action summary"] * (prompt is not None)
         res = select(["Continue", "Architecture", "Objective", "Memories", "CI"] + prompt_options, "Project Menu")
         if res == 1:
@@ -119,7 +119,7 @@ class Project:
             prompt.last_summary = get_input_from_editor(prompt.last_summary)
 
     def prompt_fields(self) -> dict:
-        from clippy.tools.architectural import templates
+        from clippinator.tools.architectural import templates
 
         default_architecture = templates['General']['architecture']
