+++ clippinator/tools/templates.yaml
