+++ clippinator/tools/terminal.py
