+++ clippinator/tools/tool.py
