+++ clippinator/tools/utils.py
